refactor: split buildAdminRouteDeps into routes_admin.go, update docs
- Move 267-line buildAdminRouteDeps to routes_admin.go
- routes_adapter.go reduced to 81 lines (registration only)
- Document deferred email resolution, remaining sideways mutations,
  legacy Server methods, and route file organization in CLAUDE.md
