+++ internal/api/routes_admin.go
@@ -1,0 +1,273 @@
+package api
+
+import (
+	"github.com/nimbleflux/fluxbase/internal/api/routes"
+)
+
+func (s *Server) buildAdminRouteDeps() *routes.AdminDeps {
+	unifiedAuth := UnifiedAuthMiddleware(s.Auth.Handler.authService, s.Auth.DashboardHandler.jwtManager, s.db)
+	return &routes.AdminDeps{
+		UnifiedAuth:        unifiedAuth,
+		RequireRole:        RequireRole,
+		TenantMiddleware:   s.Middleware.Tenant,
+		TenantDBMiddleware: s.Middleware.TenantDB,
+
+		// Subgroup dependencies
+		Branch: s.buildBranchRouteDeps(),
+		Schema: &routes.SchemaAdminDeps{
+			GetTables:               s.handleGetTables,
+			GetTableSchema:          s.handleGetTableSchema,
+			GetSchemas:              s.handleGetSchemas,
+			ExecuteQuery:            s.handleExecuteQuery,
+			ListSchemasDDL:          s.Schema.DDL.ListSchemas,
+			CreateSchemaDDL:         s.Schema.DDL.CreateSchema,
+			ListTablesDDL:           s.Schema.DDL.ListTables,
+			CreateTableDDL:          s.Schema.DDL.CreateTable,
+			DeleteTableDDL:          s.Schema.DDL.DeleteTable,
+			RenameTableDDL:          s.Schema.DDL.RenameTable,
+			AddColumnDDL:            s.Schema.DDL.AddColumn,
+			DropColumnDDL:           s.Schema.DDL.DropColumn,
+			EnableRealtime:          s.Realtime.Admin.HandleEnableRealtime,
+			ListRealtimeTables:      s.Realtime.Admin.HandleListRealtimeTables,
+			GetRealtimeStatus:       s.Realtime.Admin.HandleGetRealtimeStatus,
+			UpdateRealtimeConfig:    s.Realtime.Admin.HandleUpdateRealtimeConfig,
+			DisableRealtime:         s.Realtime.Admin.HandleDisableRealtime,
+			ExecuteSQL:              s.sqlHandler.ExecuteSQL,
+			ExportTypeScript:        s.Schema.Export.HandleExportTypeScript,
+			RefreshSchema:           s.handleRefreshSchema,
+			GetSchemaGraph:          s.GetSchemaGraph,
+			GetTableRelationships:   s.GetTableRelationships,
+			GetTablesWithRLS:        s.GetTablesWithRLS,
+			GetTableRLSStatus:       s.GetTableRLSStatus,
+			ToggleTableRLS:          s.ToggleTableRLS,
+			ListPolicies:            s.ListPolicies,
+			CreatePolicy:            s.CreatePolicy,
+			UpdatePolicy:            s.UpdatePolicy,
+			DeletePolicy:            s.DeletePolicy,
+			GetPolicyTemplates:      s.GetPolicyTemplates,
+			GetSecurityWarnings:     s.GetSecurityWarnings,
+			DumpInternalSchema:      s.Schema.InternalSchema.DumpSchema,
+			PlanInternalSchema:      s.Schema.InternalSchema.PlanSchema,
+			ApplyInternalSchema:     s.Schema.InternalSchema.ApplySchema,
+			ValidateInternalSchema:  s.Schema.InternalSchema.ValidateSchema,
+			GetInternalSchemaStatus: s.Schema.InternalSchema.GetSchemaStatus,
+			MigrateInternalSchema:   s.Schema.InternalSchema.MigrateSchema,
+		},
+		AuthProviders: &routes.AuthProvidersAdminDeps{
+			ListOAuthProviders:  s.Auth.OAuthProvider.ListOAuthProviders,
+			GetOAuthProvider:    s.Auth.OAuthProvider.GetOAuthProvider,
+			CreateOAuthProvider: s.Auth.OAuthProvider.CreateOAuthProvider,
+			UpdateOAuthProvider: s.Auth.OAuthProvider.UpdateOAuthProvider,
+			DeleteOAuthProvider: s.Auth.OAuthProvider.DeleteOAuthProvider,
+			ListSAMLProviders:   s.Auth.SAML.ListSAMLProviders,
+			GetSAMLProvider:     s.Auth.SAMLProvider.GetSAMLProvider,
+			CreateSAMLProvider:  s.Auth.SAMLProvider.CreateSAMLProvider,
+			UpdateSAMLProvider:  s.Auth.SAMLProvider.UpdateSAMLProvider,
+			DeleteSAMLProvider:  s.Auth.SAMLProvider.DeleteSAMLProvider,
+			ValidateSAML:        s.Auth.SAMLProvider.ValidateMetadata,
+			UploadSAMLMetadata:  s.Auth.SAMLProvider.UploadMetadata,
+			GetAuthSettings:     s.Auth.OAuthProvider.GetAuthSettings,
+			UpdateAuthSettings:  s.Auth.OAuthProvider.UpdateAuthSettings,
+			ListSessions:        s.Auth.AdminSession.ListSessions,
+			RevokeSession:       s.Auth.AdminSession.RevokeSession,
+			RevokeUserSessions:  s.Auth.AdminSession.RevokeUserSessions,
+		},
+		Users: &routes.UsersAdminDeps{
+			ListUsers:           s.Auth.UserManagement.ListUsers,
+			InviteUser:          s.Auth.UserManagement.InviteUser,
+			DeleteUser:          s.Auth.UserManagement.DeleteUser,
+			UpdateUser:          s.Auth.UserManagement.UpdateUser,
+			UpdateUserRole:      s.Auth.UserManagement.UpdateUserRole,
+			ResetUserPassword:   s.Auth.UserManagement.ResetUserPassword,
+			ListUsersWithQuotas: s.Quota.Handler.ListUsersWithQuotas,
+			GetUserQuota:        s.Quota.Handler.GetUserQuota,
+			SetUserQuota:        s.Quota.Handler.SetUserQuota,
+			CreateInvitation:    s.Auth.Invitation.CreateInvitation,
+			ListInvitations:     s.Auth.Invitation.ListInvitations,
+			RevokeInvitation:    s.Auth.Invitation.RevokeInvitation,
+		},
+		Tenants: &routes.TenantsAdminDeps{
+			ListMyTenants:             s.Tenancy.Tenant.ListMyTenants,
+			ListTenants:               s.Tenancy.Tenant.ListTenants,
+			ListDeletedTenants:        s.Tenancy.Tenant.ListDeletedTenants,
+			CreateTenant:              s.Tenancy.Tenant.CreateTenant,
+			GetTenant:                 s.Tenancy.Tenant.GetTenant,
+			UpdateTenant:              s.Tenancy.Tenant.UpdateTenant,
+			DeleteTenant:              s.Tenancy.Tenant.DeleteTenant,
+			RecoverTenant:             s.Tenancy.Tenant.RecoverTenant,
+			MigrateTenant:             s.Tenancy.Tenant.MigrateTenant,
+			RepairTenant:              s.Tenancy.Tenant.RepairTenant,
+			ListAdmins:                s.Tenancy.Tenant.ListAdmins,
+			AssignAdmin:               s.Tenancy.Tenant.AssignAdmin,
+			RemoveAdmin:               s.Tenancy.Tenant.RemoveAdmin,
+			GetTenantSettings:         s.Settings.Tenant.GetTenantSettings,
+			UpdateTenantSettings:      s.Settings.Tenant.UpdateTenantSettings,
+			DeleteTenantSetting:       s.Settings.Tenant.DeleteTenantSetting,
+			GetTenantSetting:          s.Settings.Tenant.GetTenantSetting,
+			GetTenantSchemaStatus:     s.Tenancy.Tenant.GetTenantSchemaStatus,
+			ApplyTenantSchema:         s.Tenancy.Tenant.ApplyTenantSchema,
+			GetStoredSchema:           s.Tenancy.Tenant.GetStoredSchema,
+			UploadTenantSchema:        s.Tenancy.Tenant.UploadTenantSchema,
+			ApplyUploadedTenantSchema: s.Tenancy.Tenant.ApplyUploadedTenantSchema,
+			DeleteStoredSchema:        s.Tenancy.Tenant.DeleteStoredSchema,
+		},
+		ServiceKeys: &routes.ServiceKeysAdminDeps{
+			ListServiceKeys:      s.Tenancy.ServiceKey.ListServiceKeys,
+			GetServiceKey:        s.Tenancy.ServiceKey.GetServiceKey,
+			CreateServiceKey:     s.Tenancy.ServiceKey.CreateServiceKey,
+			UpdateServiceKey:     s.Tenancy.ServiceKey.UpdateServiceKey,
+			DeleteServiceKey:     s.Tenancy.ServiceKey.DeleteServiceKey,
+			DisableServiceKey:    s.Tenancy.ServiceKey.DisableServiceKey,
+			EnableServiceKey:     s.Tenancy.ServiceKey.EnableServiceKey,
+			RevokeServiceKey:     s.Tenancy.ServiceKey.RevokeServiceKey,
+			DeprecateServiceKey:  s.Tenancy.ServiceKey.DeprecateServiceKey,
+			RotateServiceKey:     s.Tenancy.ServiceKey.RotateServiceKey,
+			GetRevocationHistory: s.Tenancy.ServiceKey.GetRevocationHistory,
+		},
+		Functions: &routes.FunctionsAdminDeps{
+			ReloadFunctions:        s.Functions.Handler.ReloadFunctions,
+			ListFunctionNamespaces: s.Functions.Handler.ListNamespaces,
+			ListAllExecutions:      s.Functions.Handler.ListAllExecutions,
+			GetExecutionLogs:       s.Functions.Handler.GetExecutionLogs,
+			SyncFunctions:          s.Functions.Handler.SyncFunctions,
+		},
+		Jobs: &routes.JobsAdminDeps{
+			ListJobNamespaces: s.Jobs.Handler.ListNamespaces,
+			ListJobFunctions:  s.Jobs.Handler.ListJobFunctions,
+			GetJobFunction:    s.Jobs.Handler.GetJobFunction,
+			DeleteJobFunction: s.Jobs.Handler.DeleteJobFunction,
+			GetJobStats:       s.Jobs.Handler.GetJobStats,
+			ListWorkers:       s.Jobs.Handler.ListWorkers,
+			ListAllJobs:       s.Jobs.Handler.ListAllJobs,
+			GetJobAdmin:       s.Jobs.Handler.GetJobAdmin,
+			TerminateJob:      s.Jobs.Handler.TerminateJob,
+			CancelJobAdmin:    s.Jobs.Handler.CancelJobAdmin,
+			RetryJobAdmin:     s.Jobs.Handler.RetryJobAdmin,
+			ResubmitJobAdmin:  s.Jobs.Handler.ResubmitJobAdmin,
+			SyncJobs:          s.Jobs.Handler.SyncJobs,
+		},
+		AI: &routes.AIAdminDeps{
+			ListChatbots:               s.AI.Handler.ListChatbots,
+			GetChatbot:                 s.AI.Handler.GetChatbot,
+			ToggleChatbot:              s.AI.Handler.ToggleChatbot,
+			UpdateChatbot:              s.AI.Handler.UpdateChatbot,
+			DeleteChatbot:              s.AI.Handler.DeleteChatbot,
+			SyncChatbots:               s.AI.Handler.SyncChatbots,
+			GetAIMetrics:               s.AI.Handler.GetAIMetrics,
+			ListAIProviders:            s.AI.Handler.ListProviders,
+			CreateAIProvider:           s.AI.Handler.CreateProvider,
+			UpdateAIProvider:           s.AI.Handler.UpdateProvider,
+			DeleteAIProvider:           s.AI.Handler.DeleteProvider,
+			SetDefaultAIProvider:       s.AI.Handler.SetDefaultProvider,
+			SetEmbeddingAIProvider:     s.AI.Handler.SetEmbeddingProvider,
+			ClearEmbeddingAIProvider:   s.AI.Handler.ClearEmbeddingProvider,
+			ListAIConversations:        s.AI.Handler.GetConversations,
+			GetAIConversationMessages:  s.AI.Handler.GetConversationMessages,
+			GetAIAuditLog:              s.AI.Handler.GetAuditLog,
+			ListExportableTables:       s.AI.KnowledgeBase.ListExportableTables,
+			GetExportableTableDetails:  s.AI.KnowledgeBase.GetTableDetails,
+			ExportTableToKnowledgeBase: s.AI.KnowledgeBase.ExportTableToKnowledgeBase,
+			ListKnowledgeBases:         s.AI.KnowledgeBase.ListKnowledgeBases,
+			GetKnowledgeBase:           s.AI.KnowledgeBase.GetKnowledgeBase,
+			CreateKnowledgeBase:        s.AI.KnowledgeBase.CreateKnowledgeBase,
+			UpdateKnowledgeBase:        s.AI.KnowledgeBase.UpdateKnowledgeBase,
+			ListChatbotKnowledgeBases:  s.AI.KnowledgeBase.ListChatbotKnowledgeBases,
+			LinkKnowledgeBase:          s.AI.KnowledgeBase.LinkKnowledgeBase,
+			UpdateChatbotKnowledgeBase: s.AI.KnowledgeBase.UpdateChatbotKnowledgeBase,
+			UnlinkKnowledgeBase:        s.AI.KnowledgeBase.UnlinkKnowledgeBase,
+			DeleteKnowledgeBase:        s.AI.KnowledgeBase.DeleteKnowledgeBase,
+			ListDocuments:              s.AI.KnowledgeBase.ListDocuments,
+			GetDocument:                s.AI.KnowledgeBase.GetDocument,
+			AddDocument:                s.AI.KnowledgeBase.AddDocument,
+			UploadDocument:             s.AI.KnowledgeBase.UploadDocument,
+			DeleteDocument:             s.AI.KnowledgeBase.DeleteDocument,
+			UpdateDocument:             s.AI.KnowledgeBase.UpdateDocument,
+			DeleteDocumentsByFilter:    s.AI.KnowledgeBase.DeleteDocumentsByFilter,
+			SearchKnowledgeBase:        s.AI.KnowledgeBase.SearchKnowledgeBase,
+			DebugSearch:                s.AI.KnowledgeBase.DebugSearch,
+			ListEntities:               s.AI.KnowledgeBase.ListEntities,
+			SearchEntities:             s.AI.KnowledgeBase.SearchEntities,
+			GetEntityRelationships:     s.AI.KnowledgeBase.GetEntityRelationships,
+			GetKnowledgeGraph:          s.AI.KnowledgeBase.GetKnowledgeGraph,
+			ListKBChatbots:             s.AI.KnowledgeBase.ListKnowledgeBaseChatbots,
+			CreateTableExportSync:      s.AI.KnowledgeBase.CreateTableExportSync,
+			ListTableExportSyncs:       s.AI.KnowledgeBase.ListTableExportSyncs,
+			UpdateTableExportSync:      s.AI.KnowledgeBase.UpdateTableExportSync,
+			DeleteTableExportSync:      s.AI.KnowledgeBase.DeleteTableExportSync,
+			TriggerTableExportSync:     s.AI.KnowledgeBase.TriggerTableExportSync,
+			GrantDocumentPermission:    s.AI.KnowledgeBase.GrantDocumentPermission,
+			ListDocumentPermissions:    s.AI.KnowledgeBase.ListDocumentPermissions,
+			RevokeDocumentPermission:   s.AI.KnowledgeBase.RevokeDocumentPermission,
+		},
+		RPC: &routes.RPCAdminDeps{
+			ListRPCNamespaces:   s.RPC.Handler.ListNamespaces,
+			ListProcedures:      s.RPC.Handler.ListProcedures,
+			GetProcedure:        s.RPC.Handler.GetProcedure,
+			UpdateProcedure:     s.RPC.Handler.UpdateProcedure,
+			DeleteProcedure:     s.RPC.Handler.DeleteProcedure,
+			SyncProcedures:      s.RPC.Handler.SyncProcedures,
+			ListRPCExecutions:   s.RPC.Handler.ListExecutions,
+			GetRPCExecution:     s.RPC.Handler.GetExecution,
+			GetRPCExecutionLogs: s.RPC.Handler.GetExecutionLogs,
+			CancelRPCExecution:  s.RPC.Handler.CancelExecution,
+		},
+		Logs: &routes.LogsAdminDeps{
+			ListLogs:              s.Logging.Handler.QueryLogs,
+			GetLogStats:           s.Logging.Handler.GetLogStats,
+			GetExecutionLogsAdmin: s.Logging.Handler.GetExecutionLogs,
+			FlushLogs:             s.Logging.Handler.FlushLogs,
+			GenerateTestLogs:      s.Logging.Handler.GenerateTestLogs,
+		},
+		Settings: &routes.SettingsAdminDeps{
+			ListSystemSettings:        s.Settings.System.ListSettings,
+			GetSystemSetting:          s.Settings.System.GetSetting,
+			UpdateSystemSetting:       s.Settings.System.UpdateSetting,
+			DeleteSystemSetting:       s.Settings.System.DeleteSetting,
+			CreateCustomSetting:       s.Settings.Custom.CreateSetting,
+			ListCustomSettings:        s.Settings.Custom.ListSettings,
+			CreateSecretSetting:       s.Settings.Custom.CreateSecretSetting,
+			ListSecretSettings:        s.Settings.Custom.ListSecretSettings,
+			GetSecretSetting:          s.Settings.Custom.GetSecretSetting,
+			UpdateSecretSetting:       s.Settings.Custom.UpdateSecretSetting,
+			DeleteSecretSetting:       s.Settings.Custom.DeleteSecretSetting,
+			GetUserSecretValue:        s.Settings.User.GetUserSecretValue,
+			GetCustomSetting:          s.Settings.Custom.GetSetting,
+			UpdateCustomSetting:       s.Settings.Custom.UpdateSetting,
+			DeleteCustomSetting:       s.Settings.Custom.DeleteSetting,
+			GetAppSettings:            s.Settings.App.GetAppSettings,
+			UpdateAppSettings:         s.Settings.App.UpdateAppSettings,
+			ListEmailSettings:         s.Email.Settings.GetSettings,
+			GetEmailSetting:           s.Email.Settings.GetSettings,
+			UpdateEmailSetting:        s.Email.Settings.UpdateSettings,
+			TestEmailSettings:         s.Email.Settings.TestSettings,
+			ListEmailTemplates:        s.Email.Template.ListTemplates,
+			GetEmailTemplate:          s.Email.Template.GetTemplate,
+			UpdateEmailTemplate:       s.Email.Template.UpdateTemplate,
+			TestEmailTemplate:         s.Email.Template.TestTemplate,
+			ResetEmailTemplate:        s.Email.Template.ResetTemplate,
+			GetTenantEmailSettings:    s.Email.Settings.GetSettingsForTenant,
+			UpdateTenantEmailSettings: s.Email.Settings.UpdateSettingsForTenant,
+			DeleteTenantEmailSetting:  s.Email.Settings.DeleteSettingForTenant,
+			TestTenantEmailSettings:   s.Email.Settings.TestSettingsForTenant,
+			GetCaptchaSettings:        s.Captcha.Settings.GetSettings,
+			UpdateCaptchaSettings:     s.Captcha.Settings.UpdateSettings,
+			GetInstanceSettings:       s.Settings.Instance.GetInstanceSettings,
+			UpdateInstanceSettings:    s.Settings.Instance.UpdateInstanceSettings,
+			GetOverridableSettings:    s.Settings.Instance.GetOverridableSettings,
+			UpdateOverridableSettings: s.Settings.Instance.UpdateOverridableSettings,
+		},
+		Extensions: &routes.ExtensionsAdminDeps{
+			ListExtensions:   s.Extensions.Handler.ListExtensionsForTenant,
+			GetExtension:     s.Extensions.Handler.GetExtensionStatusForTenant,
+			EnableExtension:  s.Extensions.Handler.EnableExtensionForTenant,
+			DisableExtension: s.Extensions.Handler.DisableExtensionForTenant,
+			SyncExtensions:   s.Extensions.Handler.SyncExtensions,
+		},
+		ExtensionsTenant: &routes.ExtensionsTenantDeps{
+			ListExtensions:   s.Extensions.Handler.ListExtensionsForTenant,
+			GetExtension:     s.Extensions.Handler.GetExtensionStatusForTenant,
+			EnableExtension:  s.Extensions.Handler.EnableExtensionForTenant,
+			DisableExtension: s.Extensions.Handler.DisableExtensionForTenant,
+		},
+	}
+}
